+++ lib/Braintree.php
@@ -105,6 +105,7 @@ require_once('Braintree/EqualityNode.php');
 require_once('Braintree/Exception.php');
 require_once('Braintree/Http.php');
 require_once('Braintree/KeyValueNode.php');
+require_once('Braintree/MerchantAccount.php');
 require_once('Braintree/MultipleValueNode.php');
 require_once('Braintree/MultipleValueOrTextNode.php');
 require_once('Braintree/PartialMatchNode.php');
@@ -156,6 +157,7 @@ require_once('Braintree/Xml/Generator.php');
 require_once('Braintree/Xml/Parser.php');
 require_once('Braintree/CreditCardVerification.php');
 require_once('Braintree/CreditCardVerificationSearch.php');
+require_once('Braintree/PartnerCredentials.php');
 
 if (version_compare(PHP_VERSION, '5.2.1', '<')) {
     throw new Braintree_Exception('PHP version >= 5.2.1 required');