+++ lib/Braintree/PartnerCredentials.php
@@ -1,0 +1,44 @@
+<?php
+/**
+ * Partner credentials that are created when a new partner user is created
+ *
+ * @package    Braintree
+ * @copyright  2010 Braintree Payment Solutions
+ */
+
+/**
+ * Creates an instance of PartnerCredentials as returned from a transaction
+ *
+ *
+ * @package    Braintree
+ * @copyright  2010 Braintree Payment Solutions
+ *
+ * @property-read string $merchantPublicId
+ * @property-read string $publicKey
+ * @property-read string $privateKey
+ * @property-read string $partnerUserId
+ * @uses Braintree_Instance inherits methods
+ */
+class Braintree_PartnerCredentials extends Braintree
+{
+    protected $_attributes = array();
+
+    /**
+     * @ignore
+     */
+    public static function factory($attributes)
+    {
+        $instance = new self();
+        $instance->_initialize($attributes);
+
+        return $instance;
+    }
+
+    /**
+     * @ignore
+     */
+    protected function _initialize($attributes)
+    {
+        $this->_attributes = $attributes;
+    }
+}