+++ tests/unit/WebhookNotificationTest.php
@@ -70,5 +70,84 @@ class Braintree_WebhookNotificationTest extends PHPUnit_Framework_TestCase
             $sampleNotification['payload']
         );
     }
+
+    function testBuildsASampleNotificationForAMerchantAccountApprovedWebhook()
+    {
+        $sampleNotification = Braintree_WebhookTesting::sampleNotification(
+            Braintree_WebhookNotification::SUB_MERCHANT_ACCOUNT_APPROVED,
+            "my_id"
+        );
+
+        $webhookNotification = Braintree_WebhookNotification::parse(
+            $sampleNotification['signature'],
+            $sampleNotification['payload']
+        );
+
+        $this->assertEquals(Braintree_WebhookNotification::SUB_MERCHANT_ACCOUNT_APPROVED, $webhookNotification->kind);
+        $this->assertEquals("my_id", $webhookNotification->merchantAccount->id);
+        $this->assertEquals(Braintree_MerchantAccount::STATUS_ACTIVE, $webhookNotification->merchantAccount->status);
+        $this->assertEquals("master_ma_for_my_id", $webhookNotification->merchantAccount->masterMerchantAccount->id);
+        $this->assertEquals(Braintree_MerchantAccount::STATUS_ACTIVE, $webhookNotification->merchantAccount->masterMerchantAccount->status);
+    }
+
+    function testBuildsASampleNotificationForAMerchantAccountDeclinedWebhook()
+    {
+        $sampleNotification = Braintree_WebhookTesting::sampleNotification(
+            Braintree_WebhookNotification::SUB_MERCHANT_ACCOUNT_DECLINED,
+            "my_id"
+        );
+
+        $webhookNotification = Braintree_WebhookNotification::parse(
+            $sampleNotification['signature'],
+            $sampleNotification['payload']
+        );
+
+        $this->assertEquals(Braintree_WebhookNotification::SUB_MERCHANT_ACCOUNT_DECLINED, $webhookNotification->kind);
+        $this->assertEquals("my_id", $webhookNotification->merchantAccount->id);
+        $this->assertEquals(Braintree_MerchantAccount::STATUS_SUSPENDED, $webhookNotification->merchantAccount->status);
+        $this->assertEquals("master_ma_for_my_id", $webhookNotification->merchantAccount->masterMerchantAccount->id);
+        $this->assertEquals(Braintree_MerchantAccount::STATUS_SUSPENDED, $webhookNotification->merchantAccount->masterMerchantAccount->status);
+        $this->assertEquals("Credit score is too low", $webhookNotification->message);
+        $errors = $webhookNotification->errors->forKey('merchantAccount')->onAttribute('base');
+        $this->assertEquals(Braintree_Error_Codes::MERCHANT_ACCOUNT_APPLICANT_DETAILS_DECLINED_OFAC, $errors[0]->code);
+    }
+
+    function testBuildsASampleNotificationForATransactionDisbursedWebhook()
+    {
+        $sampleNotification = Braintree_WebhookTesting::sampleNotification(
+            Braintree_WebhookNotification::TRANSACTION_DISBURSED,
+            "my_id"
+        );
+
+        $webhookNotification = Braintree_WebhookNotification::parse(
+            $sampleNotification['signature'],
+            $sampleNotification['payload']
+        );
+
+        $this->assertEquals(Braintree_WebhookNotification::TRANSACTION_DISBURSED, $webhookNotification->kind);
+        $this->assertEquals("my_id", $webhookNotification->transaction->id);
+        $this->assertEquals(100, $webhookNotification->transaction->amount);
+        $this->assertNotNull($webhookNotification->transaction->disbursementDetails->disbursementDate);
+    }
+
+    function testBuildsASampleNotificationForAPartnerUserCreatedWebhook()
+    {
+        $sampleNotification = Braintree_WebhookTesting::sampleNotification(
+            Braintree_WebhookNotification::PARTNER_USER_CREATED,
+            "my_id"
+        );
+
+        $webhookNotification = Braintree_WebhookNotification::parse(
+            $sampleNotification['signature'],
+            $sampleNotification['payload']
+        );
+
+        $this->assertEquals(Braintree_WebhookNotification::PARTNER_USER_CREATED, $webhookNotification->kind);
+        $this->assertEquals("my_id", $webhookNotification->partnerCredentials->id);
+        $this->assertEquals("public_id", $webhookNotification->partnerCredentials->merchantPublicId);
+        $this->assertEquals("public_key", $webhookNotification->partnerCredentials->publicKey);
+        $this->assertEquals("private_key", $webhookNotification->partnerCredentials->privateKey);
+        $this->assertEquals("abc123", $webhookNotification->partnerCredentials->partnerUserId);
+    }
 }
 ?>